requirements.txt
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
idna==3.4
markdown2==2.4.6
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
Pygments==2.13.0
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
svgwrite==1.4.3
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.12
wavedrom==2.0.3.post3
